Fix node version MTA build
The MTA build (currently for the Node.js version only) comes up with a build profile for the Java stack by default, which does not work of course. This change defines build settings in package.json for the Node stack explicitly, and selects the corresponding profile in the mta.yaml file.